requirements.txt
streamlit
google-genai
pandas
python-multipart
graphviz
plotly
watchdog
